+++ src/messages/en.json
@@ -1,4 +1,8 @@
 {
+  "Metadata": {
+    "title": "Polio — Sukhyeon's Portfolio",
+    "description": "Portfolio of Sukhyeon, a frontend engineer who solves hard problems with structure and builds systems that didn't exist before."
+  },
   "Header": {
     "home": "Home"
   },
