+++ src/messages/ko.json
@@ -1,4 +1,8 @@
 {
+  "Metadata": {
+    "title": "Polio — 숙현의 포트폴리오",
+    "description": "복잡한 문제를 구조로 풀고, 없던 체계를 만드는 프론트엔드 엔지니어 숙현의 포트폴리오입니다."
+  },
   "Header": {
     "home": "홈"
   },
